@@ -1,8 +1,17 @@
 module github.com/Gonnekone/proto-contract
 
-go 1.21.3
+go 1.22
+
+toolchain go1.22.1
 
 require (
-	google.golang.org/grpc/cmd/protoc-gen-go-grpc v1.2.0 // indirect
-	google.golang.org/protobuf v1.28.1 // indirect
+	google.golang.org/grpc v1.69.4
+	google.golang.org/protobuf v1.35.1
+)
+
+require (
+	golang.org/x/net v0.30.0 // indirect
+	golang.org/x/sys v0.26.0 // indirect
+	golang.org/x/text v0.19.0 // indirect
+	google.golang.org/genproto/googleapis/rpc v0.0.0-20241015192408-796eee8c2d53 // indirect
 )
